requirements.txt
aiogram==2.25.2
openai==0.28.0
python-dotenv==0.21.1
